requirements.txt
requests
python-dotenv
google-cloud-bigquery
functions-framework>=3.0.0
